Added tag for specifying which decode strategy to use for values. This ultimately allows defining values with YAML syntax and having them decoded into Slice and Map.
Tag is of the the form `decode=<value>`

Support values are:
* type
* kind
* yaml

YAML was chosen because it is a superset of JSON and thus both will parse with the YAML parser. Additionally YAML does not require the explicit quotation of keys and is simpler to write inline strings with. Both represent probably the best ways to define a list / map with inline strings. YAML is not included in the golang:encode package, but the go-yaml project is heavily supported and used.

By default / omission, the Type strategy will be used and then Kind for fallback if that fails. This allows lazy defining the tags and leaving the decode strategy off and if there is a Type defined for decoding it will use it, else use the Kind of the var.

Slices and Maps have been added to the Kind decode strategy. They use their own function which then calls the YAML parse.

Each decoding strategy uses its own function. This keeps each to their single responsibility and eliminates the ambiguity of using multiple decision values in a conversion (mixing Kind and Type).